requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
distlib==0.3.7
Django==4.2.3
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-filter==23.2
django-tinymce==3.6.1
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.2
idna==3.4
MouseInfo==0.1.3
mss==9.0.1
Pillow==9.5.0
platformdirs==3.9.1
psycopg2==2.9.7
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sslcommerz-lib==1.0
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.2
virtualenvwrapper-win==1.2.7
